requirements.txt
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
sqlalchemy>=2.0.30
python-dotenv>=1.0.1
firebase-admin>=6.5.0
python-multipart>=0.0.9
google-generativeai>=0.8.0
